// frontend/src/app/models/stavka-korpe.ts
export class StavkaKorpe {
    sifra: string = "";
    naziv: string = "";
    kategorija: string = "";
    kolicina: number = 1;
    tipStampe: string = "";
    cenaPoKomadu: number = 0;
    stamparijaKorisnickoIme: string = "";
    nazivStamparije: string = "";
    boja: string = "";
    personalizacijaTekst: string = "";
    personalizacijaSlika: string = "";
}